+++ src/main/java/com/notenoughmail/kubejs_tfc/block/sub/TFCPathBlockBuilder.java
@@ -1,16 +1,15 @@
-package com.notenoughmail.kubejs_tfc.block.internal;
+package com.notenoughmail.kubejs_tfc.block.sub;
 
 import com.notenoughmail.kubejs_tfc.block.TFCDirtBlockBuilder;
 import dev.latvian.mods.kubejs.block.BlockBuilder;
 import dev.latvian.mods.kubejs.block.BlockItemBuilder;
+import dev.latvian.mods.kubejs.client.VariantBlockStateGenerator;
 import dev.latvian.mods.kubejs.generator.AssetJsonGenerator;
 import dev.latvian.mods.kubejs.generator.DataJsonGenerator;
 import dev.latvian.mods.kubejs.loot.LootBuilder;
 import dev.latvian.mods.kubejs.typings.Generics;
-import net.dries007.tfc.common.TFCTags;
-import net.dries007.tfc.common.blockentities.TFCBlockEntities;
-import net.dries007.tfc.common.blocks.ExtendedProperties;
-import net.dries007.tfc.common.blocks.soil.FarmlandBlock;
+import dev.latvian.mods.kubejs.typings.Info;
+import net.dries007.tfc.common.blocks.soil.PathBlock;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
@@ -18,20 +17,26 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
-@SuppressWarnings("unused")
-public class TFCFarmlandBlockBuilder extends ExtendPropertiesBlockBuilder {
+public class TFCPathBlockBuilder extends BlockBuilder {
 
     public transient final TFCDirtBlockBuilder parent;
+    public transient boolean useUniqueDirtTexture;
 
-    public TFCFarmlandBlockBuilder(ResourceLocation i, TFCDirtBlockBuilder parent) {
+    public TFCPathBlockBuilder(ResourceLocation i, TFCDirtBlockBuilder parent) {
         super(i);
         this.parent = parent;
-        tag(TFCTags.Blocks.FARMLAND.location());
+        useUniqueDirtTexture = false;
+    }
+
+    @Info(value = "Makes the path block use a unique texture for the dirt part of its texture, by default uses the texture of its parent dirt block")
+    public TFCPathBlockBuilder uniqueDirtTexture() {
+        useUniqueDirtTexture = true;
+        return this;
     }
 
     @Override
     public Block createObject() {
-        return new FarmlandBlock(createExtendedProperties(), parent);
+        return new PathBlock(createProperties(), parent);
     }
 
     @Override
@@ -64,17 +69,26 @@ public class TFCFarmlandBlockBuilder extends ExtendPropertiesBlockBuilder {
     }
 
     @Override
-    public ExtendedProperties createExtendedProperties() {
-        return super.createExtendedProperties()
-                .blockEntity(TFCBlockEntities.FARMLAND);
+    protected void generateBlockModelJsons(AssetJsonGenerator generator) {
+        final String base = newID("block/", "").toString();
+
+        generator.blockModel(id, m -> {
+            m.parent("tfc:block/grass_path");
+            m.texture("dirt", useUniqueDirtTexture ? base : parent.newID("block/", "").toString());
+            m.texture("top", base + "_top");
+            m.texture("side", base + "_side");
+        });
     }
 
     @Override
-    protected void generateBlockModelJsons(AssetJsonGenerator generator) {
-        generator.blockModel(id, m -> {
-            m.parent("block/template_farmland");
-            m.texture("top", newID("block/", "").toString());
-            m.texture("dirt", parent.textures.get("particle").getAsString());
+    protected void generateBlockStateJson(VariantBlockStateGenerator bs) {
+        final String model = newID("block/", "").toString();
+
+        bs.variant("", v -> {
+            v.model(model);
+            v.model(model).y(90);
+            v.model(model).y(180);
+            v.model(model).y(270);
         });
     }
 }