+++ src/main/java/com/notenoughmail/kubejs_tfc/block/sub/WindmillBlockBuilder.java
@@ -1,0 +1,70 @@
+package com.notenoughmail.kubejs_tfc.block.sub;
+
+import com.notenoughmail.kubejs_tfc.block.AxleBlockBuilder;
+import com.notenoughmail.kubejs_tfc.block.internal.ExtendedPropertiesBlockBuilder;
+import com.notenoughmail.kubejs_tfc.util.RegistryUtils;
+import dev.latvian.mods.kubejs.client.ModelGenerator;
+import dev.latvian.mods.kubejs.client.VariantBlockStateGenerator;
+import dev.latvian.mods.kubejs.generator.AssetJsonGenerator;
+import dev.latvian.mods.kubejs.generator.DataJsonGenerator;
+import dev.latvian.mods.kubejs.loot.LootBuilder;
+import dev.latvian.mods.kubejs.util.UtilsJS;
+import net.dries007.tfc.common.blockentities.TFCBlockEntities;
+import net.dries007.tfc.common.blockentities.rotation.WindmillBlockEntity;
+import net.dries007.tfc.common.blocks.ExtendedProperties;
+import net.dries007.tfc.common.blocks.rotation.WindmillBlock;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.block.Block;
+
+public class WindmillBlockBuilder extends ExtendedPropertiesBlockBuilder {
+
+    public transient final AxleBlockBuilder parent;
+
+    public WindmillBlockBuilder(ResourceLocation i, AxleBlockBuilder parent) {
+        super(i);
+        this.parent = parent;
+        noItem();
+        RegistryUtils.hackBlockEntity(TFCBlockEntities.WINDMILL, this);
+    }
+
+    @Override
+    public Block createObject() {
+        return new WindmillBlock(createExtendedProperties(), UtilsJS.cast(parent));
+    }
+
+    @Override
+    public ExtendedProperties createExtendedProperties() {
+        return super.createExtendedProperties()
+                .noOcclusion()
+                .blockEntity(TFCBlockEntities.WINDMILL)
+                .ticks(WindmillBlockEntity::serverTick, WindmillBlockEntity::clientTick);
+    }
+
+    @Override
+    public void generateDataJsons(DataJsonGenerator generator) {
+        var lootBuilder = new LootBuilder(null);
+        lootBuilder.type = "minecraft:block";
+
+        if (lootTable != null) {
+            lootTable.accept(lootBuilder);
+        } else if (parent.get().asItem() != Items.AIR) {
+            lootBuilder.addPool(pool -> {
+                pool.survivesExplosion();
+                pool.addItem(new ItemStack(parent.get()));
+            });
+        }
+
+        generator.json(newID("loot_tables/blocks/", ""), lootBuilder.toJson());
+    }
+
+    @Override
+    protected void generateBlockModelJsons(AssetJsonGenerator generator) {
+    }
+
+    @Override
+    protected void generateBlockStateJson(VariantBlockStateGenerator bs) {
+        bs.simpleVariant("", "tfc:block/empty");
+    }
+}